OMAPI/TMAPI: Set MSS on es10x commands to 0xFF for channels
https://android.googlesource.com/platform/frameworks/opt/telephony/+/master/src/java/com/android/internal/telephony/uicc/euicc/apdu/RequestBuilder.java#31

+++ libs/lpac-jni/src/main/java/net/typeblog/lpac_jni/LpacJni.kt
@@ -14,7 +14,7 @@ internal object LpacJni {
     external fun destroyContext(handle: Long)
 
     external fun euiccInit(handle: Long): Int
-    external fun euiccSetMss(handle: Long, mss: Byte)
+    external fun euiccSetMss(handle: Long, mss: Int)
     external fun euiccFini(handle: Long)
 
     // es10c
