+++ libs/lpac-jni/src/main/java/net/typeblog/lpac_jni/impl/LocalProfileAssistantImpl.kt
@@ -100,8 +100,8 @@ class LocalProfileAssistantImpl(
         httpInterface.usePublicKeyIds(pkids.toTypedArray())
     }
 
-    override fun setEs10xMss(mss: Byte) {
-        LpacJni.euiccSetMss(contextHandle, mss)
+    override fun setEs10xMss(mss: UByte) {
+        LpacJni.euiccSetMss(contextHandle, mss.toInt())
     }
 
     override val valid: Boolean
